Separate DTOs to requests and responses.
