+++ CityInfo.API/Models/Requests/PointOfInterestForUpdateDto.cs
@@ -1,9 +1,9 @@
-namespace CityInfo.API.Models;
+namespace CityInfo.API.Models.Requests;
 
-public class PointOfInterestDto
+public class PointOfInterestForUpdateDto
 {
     public int Id { get; set; }
-
+    
     public string Name { get; set; } = string.Empty;
     
     public string? Description { get; set; }
