+++ CityInfo.API/Models/Responses/PointOfInterestDto.cs
@@ -1,9 +1,9 @@
-namespace CityInfo.API.Models;
+namespace CityInfo.API.Models.Responses;
 
-public class PointOfInterestForUpdateDto
+public class PointOfInterestDto
 {
     public int Id { get; set; }
-    
+
     public string Name { get; set; } = string.Empty;
     
     public string? Description { get; set; }
